+++ sample-apps/react-native/dogfood/src/components/NavigationHeader.tsx
@@ -16,6 +16,7 @@ import { appTheme } from '../theme';
 import { AVATAR_SIZE } from '../constants';
 import { Button } from './Button';
 import { ButtonTestIds } from '../constants/TestIds';
+import { REACT_NATIVE_DOGFOOD_APP_ENVIRONMENT } from '@env';
 
 export const NavigationHeader = ({ route }: NativeStackHeaderProps) => {
   const videoClient = useStreamVideoClient();
@@ -59,10 +60,11 @@ export const NavigationHeader = ({ route }: NativeStackHeaderProps) => {
   };
 
   const showChooseModeButton =
-    route.name === 'JoinMeetingScreen' ||
-    route.name === 'JoinCallScreen' ||
-    route.name === 'AudioRoom' ||
-    route.name === 'LiveStreamChoose';
+    REACT_NATIVE_DOGFOOD_APP_ENVIRONMENT === 'pronto' &&
+    (route.name === 'JoinMeetingScreen' ||
+      route.name === 'JoinCallScreen' ||
+      route.name === 'AudioRoom' ||
+      route.name === 'LiveStreamChoose');
 
   return (
     <SafeAreaView style={styles.header} edges={['top']}>