+++ sample-apps/react-native/dogfood/src/types/env.d.ts
@@ -1,0 +1,5 @@
+type AppEnvironment = 'pronto' | 'demo';
+
+declare module '@env' {
+  export const REACT_NATIVE_DOGFOOD_APP_ENVIRONMENT: AppEnvironment;
+}